M153/requirements.txt
psycopg2-binary==2.9.3
rich==11.0.0